Relocate build_stats helper to actor_loop_flow
Move the 81-LOC `build_stats` free fn out of turn.rs into
actor_loop_flow.rs. Its only consumer was the already-migrated
run_actor_loop, so this also removes the `use super::turn::build_stats`
back-import from actor_loop_flow.rs.

Imports added to actor_loop_flow.rs:

- std::collections::HashSet
- std::path::Path
- crate::util::workspace_paths::is_ignored_workspace_display_path
- super::summary::LoopStats

Part of #680 / #681 — closes the loop on run_actor_loop's biggest
back-import dependency. After this, turn.rs is ~81 LOC closer to
the Phase 1 receive condition (turn.rs ≤ 35,000).
